Move type api examples into a test case.
