+++ graphql-api.cabal
@@ -27,13 +27,13 @@ library
   build-depends:
       base >= 4.9 && < 5
     , protolude
+    , exceptions
+    , transformers
+    , attoparsec
     , aeson
     , containers
     , graphql
     , text
-    , attoparsec
-    , transformers
-    , exceptions
   exposed-modules:
       GraphQL
       GraphQL.API
@@ -56,6 +56,9 @@ executable graphql-api-example
   build-depends:
       base >= 4.9 && < 5
     , protolude
+    , exceptions
+    , transformers
+    , attoparsec
     , graphql-api
   default-language: Haskell2010
 
@@ -69,6 +72,9 @@ test-suite graphql-api-tests
   build-depends:
       base >= 4.9 && < 5
     , protolude
+    , exceptions
+    , transformers
+    , attoparsec
     , aeson
     , containers
     , graphql
@@ -76,5 +82,6 @@ test-suite graphql-api-tests
     , tasty
     , tasty-hspec
   other-modules:
+      TypeApiTests
       TypeTests
   default-language: Haskell2010
